+++ include/indexes/full_scan.hpp
@@ -25,7 +25,7 @@ public:
     static std::unique_ptr<Table> scan_partition(
             Table *t,
             Query& query,
-            std::vector<std::pair<int64_t, int64_t> >& partitions,
+            std::vector<std::pair<size_t, size_t> >& partitions,
             std::vector<bool>& partition_skip
             );
 };
