+++ src/data/generators/uniform_generator.cpp
@@ -20,9 +20,9 @@ void UniformGenerator::generate(std::string table_path, std::string query_path){
     std::mt19937 generator(rand_dev());
     std::uniform_int_distribution<int> distr(0, n_rows);
 
-    for(int64_t i = 0; i < n_rows; ++i){
+    for(size_t i = 0; i < n_rows; ++i){
         std::vector<float> row(n_dimensions);
-        for(int64_t j = 0; j < n_dimensions; ++j){
+        for(size_t j = 0; j < n_dimensions; ++j){
            row.at(j) = distr(generator); 
         }
         table->append(&(row[0]));
@@ -39,12 +39,12 @@ void UniformGenerator::generate(std::string table_path, std::string query_path){
         0, n_rows*(1-per_column_selectivity)
     );
 
-    for(int64_t i = 0; i < n_queries; ++i){
+    for(size_t i = 0; i < n_queries; ++i){
         std::vector<float> lows(n_dimensions);
         std::vector<float> highs(n_dimensions);
-        std::vector<int64_t> cols(n_dimensions);
+        std::vector<size_t> cols(n_dimensions);
 
-        for(int64_t j = 0; j < n_dimensions; ++j){
+        for(size_t j = 0; j < n_dimensions; ++j){
             lows.at(j) = distr_query(generator_query);
             highs.at(j) = lows.at(j) + (int64_t)(n_rows * per_column_selectivity);
             cols.at(j) = j;
